test: add tests for the frame merge algorithm
